refactor(api)!: retire suggest_config / describe_analysis_modes (#448)
suggest_config / describe_analysis_modes / SuggestConfigResult exit
the public surface; inspect_panel (#443) covers the same pre-flight
ground with a typed PanelInspection, and the new evaluate signature
no longer raises MissingConfigError via the agent-recovery flow that
suggest_config was the named recovery path for.

list_metrics moves to factrix/_metric_index.py and list_estimators to
factrix/stats/__init__.py — both alongside the registries they read
so no third module survives the deletion of _describe.py. The two
internal panel-shape detectors (_detect_signal / _detect_scope) move
into factrix/_inspect.py, their only surviving caller.

mkdocs.yml / docs/api/index.md / docs/api/errors.md /
docs/api/analysis-config.md drop the dedicated suggest-config nav
entry and dead inbound links; remaining stale prose mentioning the
retired helpers stays for the C6 editorial sweep.

Refs #448.
